* 'master' of https://github.com/WarHead/uwom-server:
  Core: Sehr viele Kompilierungswarnungen gefixt, und überflüssigen Debugcode aus dem AHBot entfernt.
  Jail: Kompilierungswarnungen gefixt, und einige Tweaks.

Conflicts:
	src/server/game/World/World.h
	src/server/scripts/Northrend/tausendwinter.cpp
	src/server/scripts/OutdoorPvP/OutdoorPvPTW.cpp
	src/server/scripts/OutdoorPvP/OutdoorPvPTW.h

Konflikte gelöst.

+++ src/server/game/Spells/SpellInfo.cpp
@@ -258,7 +258,7 @@ bool SpellEffectInfo::IsEffect() const
 
 bool SpellEffectInfo::IsEffect(SpellEffects effectName) const
 {
-    return Effect == effectName;
+    return Effect == uint32(effectName);
 }
 
 bool SpellEffectInfo::IsAura() const
@@ -268,7 +268,7 @@ bool SpellEffectInfo::IsAura() const
 
 bool SpellEffectInfo::IsAura(AuraType aura) const
 {
-    return IsAura() && ApplyAuraName == aura;
+    return IsAura() && ApplyAuraName == uint32(aura);
 }
 
 bool SpellEffectInfo::IsArea() const
